Force rename DTO to Dtos
