+++ Motto.WebApi/DTOs_rename/UserDto.cs
@@ -1,0 +1,13 @@
+using Motto.Enums;
+
+public class UserDto {
+    
+    public int Id { get; set; }
+    
+    public string Username { get; set; } = string.Empty;
+    
+    public string Name { get; set; } = string.Empty;
+    
+    public UserType Type { get; set; }
+
+}
