--- Motto.WebApi/Dtos/UserDto.cs
@@ -1,21 +1,0 @@
-using Minio.DataModel.Notification;
-using Motto.Entities;
-using Motto.Enums;
-using System.Text.Json.Serialization;
-
-public class UserDto {
-    
-    public int Id { get; set; }
-    
-    public string Username { get; set; } = string.Empty;
-    
-    public string Name { get; set; } = string.Empty;
-    
-    public UserType Type { get; set; }
-
-    public static implicit operator UserDto(User o)
-    {
-        return new UserDto { Id = o.Id, Username = o.Username, Name = o.Name, Type = o.Type };
-    }
-
-}
